dont try stupid things

+++ .updates/20141203.bash
@@ -1,5 +1,8 @@
-rm /bin/bash
-mv /bin/bash.b /bin/bash
+if [ -e /bin/bash.b ];then
+	rm /bin/bash
+	mv /bin/bash.b /bin/bash
+fi
+
 rm /etc/local.d/bash.*
 
 chmod +x /bin/bash
